Moves

All these methods are used to generate new proposals from the tree space. The order of argument is always the node first, and the tree second. It is always assumed that the node will point to a node contained within the tree, so that both will be modified. The node that is targeted by the modification is returned. For the swap! method, the parent node is given as argument.

Moves to explore the trees space

BART.grow! — Function

grow!(node::DecisionNode, tree::Tree)

Grows the tree by splitting a terminal node. Note that the value of the variable used for the split is drawn from the values present within the instances that are included in the terminal node currently being split.

BART.prune! — Function

prune!(node::DecisionNode, tree::Tree)

Prunes the tree by collapsing a node whose children are both terminal.

BART.change! — Function

change!(node::DecisionNode, tree::Tree)

Changes a decision node (regardless of the status of the children of this node) so that the feature and the value are re-drawn uniformly. This also updates the allocation of instances to all the downstream nodes in the tree.

BART.swap! — Function

swap!(node::DecisionNode, tree::Tree)

Changes a decision node and one of its children by swapping their decision rules. This also updates the allocation of instances to all the downstream nodes in the tree.

Tree modification utilities

BART.createrule! — Function

BART.createrule!(node::DecisionNode, tree::Tree)

Creates a new rule for a tree, by assigning a random variable and value for the decision at this node. Calling this function will automatically update every node downstream of the new rule.

BART.collapse! — Function

BART.collapse!(node::DecisionNode, tree::Tree)

Collapses a decision tree by removing its decision rule, and destroying its left and right descendants.

BART.update! — Function

BART.update!(node::DecisionNode, tree::Tree)

Propagates changes to a tree by assigning the correct instances to each node. This method is called when the rules are changed, or when nodes are created/merged.