Trees

Overview of the types

BART.Tree — Type

Tree

A tree is a simple datastructure that binds a response y, a matrix of features X, and the tree itself stored in root. When the type is constructed without a root, the default behavior is to create a root that is a terminal node of depth 0, with a response set to the average of y.

BART.DecisionNode — Type

DecisionNode{T}

A decision node that works for both classification and regression.

Tree utilities

BART.depth — Function

depth(node::DecisionNode)

Returns the depth of a node, which is stored in the node metadata.

depth(tree::Tree)

Returns the depth of a tree, which is measured as the lower depth of its terminal nodes.